remove unnecessary host class for single message router
