+++ src/host_beta/mainHost.ts
@@ -1,5 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { Host } from './host';
-export let host = new Host(window as unknown as Window);
+import { MessageRouter } from './messageRouter';
+export let messageRouter = new MessageRouter(window as unknown as Window);
